requirements.txt
asgiref==3.2.10
Django==3.1
djangorestframework==3.11.1
factory-boy==2.12.0
Faker==4.1.1
python-dateutil==2.8.1
pytz==2020.1
six==1.15.0
sqlparse==0.3.1
text-unidecode==1.3
